Bluetooth Terminal Activity is removed.
Due to error of Activity lifecycle removing BluetoothTerm
