+++ app/src/main/AndroidManifest.xml
@@ -20,7 +20,6 @@
             </intent-filter>
         </activity>
         <activity android:name=".Dashboard" />
-        <activity android:name=".bluetooth_connectivity.BluetoothTerm"/>
         <activity android:name=".bluetooth_connectivity.DeviceList"/>
     </application>
