--- app/src/main/java/com/example/ravi/mayosa/bluetooth_connectivity/BluetoothTerm.java
@@ -1,297 +1,0 @@
-package com.example.ravi.mayosa.bluetooth_connectivity;
-
-
-import android.app.Activity;
-import android.bluetooth.BluetoothAdapter;
-import android.bluetooth.BluetoothDevice;
-import android.content.Intent;
-import android.os.Bundle;
-import android.os.Handler;
-import android.os.Message;
-import android.support.v7.widget.DefaultItemAnimator;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.util.Log;
-import android.view.KeyEvent;
-import android.view.View;
-import android.view.View.OnClickListener;
-import android.view.inputmethod.EditorInfo;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.TextView;
-import android.widget.Toast;
-
-import com.example.ravi.mayosa.Database.DataRecord;
-import com.example.ravi.mayosa.HomeActivity;
-import com.example.ravi.mayosa.R;
-import com.example.ravi.mayosa.valueSeque;
-
-import java.util.ArrayList;
-import java.util.List;
-
-
-public class BluetoothTerm extends Activity {
-
-    // Message types sent from the BluetoothComService Handler
-    public static final int MESSAGE_STATE_CHANGE = 1;
-    public static final int MESSAGE_READ = 2;
-    public static final int MESSAGE_WRITE = 3;
-    public static final int MESSAGE_DEVICE_NAME = 4;
-    public static final int MESSAGE_TOAST = 5;
-    private String to_store = "";
-
-
-    // Key names received from the BluetoothComService Handler
-    public static final String DEVICE_NAME = "device_name";
-    public static final String TOAST = "toast";
-
-    // Intent request codes
-    private static final int REQUEST_CONNECT_DEVICE = 1;
-    private static final int REQUEST_ENABLE_BT = 2;
-    private EditText mOutEditText;
-    private Button mSendButton;
-
-    // Name of the connected device
-    private String mConnectedDeviceName = null;
-    // String buffer for outgoing messages
-    private StringBuffer mOutStringBuffer;
-
-    // Local Bluetooth adapter
-    private BluetoothAdapter mBluetoothAdapter = null;
-
-    // Member object for the chat services
-    private BluetoothComService mChatService = null;
-
-    private RecyclerView mRecyclerView;
-    private LinearLayoutManager mLayoutManager;
-    private MessageAdapter mAdapter;
-    private LinearLayoutManager linearLayoutManager;
-    public int counter = 0;
-
-    private List<com.example.ravi.mayosa.bluetooth_connectivity.Message> messageList = new ArrayList<com.example.ravi.mayosa.bluetooth_connectivity.Message>();
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-        setContentView(R.layout.activity_bluetooth_term);
-
-        mRecyclerView = (RecyclerView) findViewById(R.id.my_recycler_view);
-        mRecyclerView.setHasFixedSize(true);
-        mLayoutManager = new LinearLayoutManager(this);
-        mRecyclerView.setLayoutManager(mLayoutManager);
-        mAdapter = new MessageAdapter(getBaseContext(), messageList);
-        mRecyclerView.setAdapter(mAdapter);
-        mRecyclerView.setItemAnimator(new DefaultItemAnimator());
-        linearLayoutManager = new LinearLayoutManager(getApplicationContext());
-        linearLayoutManager.scrollToPosition(mAdapter.getItemCount());
-        mRecyclerView.setLayoutManager(linearLayoutManager);
-        mBluetoothAdapter = BluetoothAdapter.getDefaultAdapter();
-
-        // If the adapter is null, then Bluetooth is not supported
-        if (mBluetoothAdapter == null) {
-            Toast.makeText(this, "Bluetooth is not available", Toast.LENGTH_LONG).show();
-            finish();
-            return;
-        }
-    }
-
-    @Override
-    public void onStart() {
-        super.onStart();
-        if (!mBluetoothAdapter.isEnabled()) {
-            Intent enableIntent = new Intent(BluetoothAdapter.ACTION_REQUEST_ENABLE);
-            startActivityForResult(enableIntent, REQUEST_ENABLE_BT);
-        } else {
-            if (mChatService == null) setupChat();
-        }
-    }
-
-    @Override
-    public synchronized void onResume() {
-        super.onResume();
-        if (mChatService != null) {
-            if (mChatService.getState() == BluetoothComService.STATE_NONE) {
-                mChatService.start();
-            }
-        }
-    }
-
-    private void setupChat() {
-        mOutEditText = (EditText) findViewById(R.id.edit_text_out);
-        mOutEditText.setOnEditorActionListener(mWriteListener);
-        mSendButton = (Button) findViewById(R.id.button_send);
-        mSendButton.setOnClickListener(new OnClickListener() {
-            public void onClick(View v) {
-                TextView view = (TextView) findViewById(R.id.edit_text_out);
-                String message = view.getText().toString();
-                sendMessage(message);
-            }
-        });
-
-        // Initialize the BluetoothComService to perform bluetooth connections
-        mChatService = new BluetoothComService(this, mHandler);
-
-        // Initialize the buffer for outgoing messages
-        mOutStringBuffer = new StringBuffer("");
-    }
-
-    @Override
-    public synchronized void onPause() {
-        super.onPause();
-
-    }
-
-    @Override
-    public void onStop() {
-        super.onStop();
-
-    }
-
-    @Override
-    public void onDestroy() {
-        super.onDestroy();
-        // Stop the Bluetooth chat services
-        if (mChatService != null) mChatService.stop();
-    }
-
-    private void ensureDiscoverable() {
-        if (mBluetoothAdapter.getScanMode() !=
-                BluetoothAdapter.SCAN_MODE_CONNECTABLE_DISCOVERABLE) {
-            Intent discoverableIntent = new Intent(BluetoothAdapter.ACTION_REQUEST_DISCOVERABLE);
-            discoverableIntent.putExtra(BluetoothAdapter.EXTRA_DISCOVERABLE_DURATION, 300);
-            startActivity(discoverableIntent);
-        }
-    }
-
-
-    private void sendMessage(String message) {
-
-        // Check that we're actually connected before trying anything
-        if (mChatService.getState() != BluetoothComService.STATE_CONNECTED) {
-            Toast.makeText(this, R.string.not_connected, Toast.LENGTH_SHORT).show();
-            return;
-        }
-        // Check that there's actually something to send
-        if (message.length() > 0) {
-            // Get the message bytes and tell the BluetoothComService to write
-            byte[] send = message.getBytes();
-            mChatService.write(send);
-            // Reset out string buffer to zero and clear the edit text field
-            mOutStringBuffer.setLength(0);
-            mOutEditText.setText(mOutStringBuffer);
-        }
-    }
-
-    // The action listener for the EditText widget, to listen for the return key
-    private TextView.OnEditorActionListener mWriteListener =
-            new TextView.OnEditorActionListener() {
-                public boolean onEditorAction(TextView view, int actionId, KeyEvent event) {
-                    // If the action is a key-up event on the return key, send the message
-                    if (actionId == EditorInfo.IME_NULL && event.getAction() == KeyEvent.ACTION_UP) {
-                        String message = view.getText().toString();
-                        sendMessage(message);
-                    }
-                    return true;
-                }
-            };
-
-    // The Handler that gets information back from the BluetoothComService
-    private final Handler mHandler = new Handler() {
-        @Override
-        public void handleMessage(Message msg) {
-            switch (msg.what) {
-                case MESSAGE_WRITE:
-                    byte[] writeBuf = (byte[]) msg.obj;
-                    // construct a string from the buffer
-                    String writeMessage = new String(writeBuf);
-                    mAdapter.notifyDataSetChanged();
-                    linearLayoutManager.scrollToPosition(mAdapter.getItemCount());
-                    messageList.add(new com.example.ravi.mayosa.bluetooth_connectivity.Message(counter++, writeMessage, "Me"));
-                    break;
-                case MESSAGE_READ:
-                    String readMessage =(String) msg.obj;
-                    // construct a string from the valid bytes in the buffer
-                    //String readMessage = new String(readBuf, 0, msg.arg1);
-
-                    if(to_store.split(",").length>13){
-                        to_store="";
-                        break;
-                    }
-
-                    if (to_store.length()!=0){
-                        to_store = to_store+readMessage;
-                    }
-                    else{
-                        to_store = readMessage;
-                    }
-
-                    Log.e("Message read","*****************     "+to_store+"  "+to_store.split(",").length);
-
-
-                    if ((to_store.split(",").length==13)){
-                        Log.e("Splited", to_store);
-                        DataRecord tempRec = new DataRecord(to_store);
-                        valueSeque.addRecord(tempRec);
-                        mAdapter.notifyDataSetChanged();
-                        linearLayoutManager.scrollToPosition(mAdapter.getItemCount());
-                        messageList.add(new com.example.ravi.mayosa.bluetooth_connectivity.Message(counter++, to_store, mConnectedDeviceName));
-                        to_store = "";
-                    }
-
-                    break;
-                case MESSAGE_DEVICE_NAME:
-                    // save the connected device's name
-                    mConnectedDeviceName = msg.getData().getString(DEVICE_NAME);
-                    Toast.makeText(getApplicationContext(), "Connected to " + mConnectedDeviceName, Toast.LENGTH_SHORT).show();
-                    break;
-                case MESSAGE_TOAST:
-                    Toast.makeText(getApplicationContext(), msg.getData().getString(TOAST),
-                            Toast.LENGTH_SHORT).show();
-                    break;
-            }
-        }
-    };
-
-    public void onActivityResult(int requestCode, int resultCode, Intent data) {
-        switch (requestCode) {
-            case REQUEST_CONNECT_DEVICE:
-                // When DeviceListActivity returns with a device to connect
-                if (resultCode == Activity.RESULT_OK) {
-                    // Get the device MAC address
-                    String address = data.getExtras().getString(DeviceList.EXTRA_DEVICE_ADDRESS);
-                    // Get the BLuetoothDevice object
-                    BluetoothDevice device = mBluetoothAdapter.getRemoteDevice(address);
-                    // Attempt to connect to the device
-                    mChatService.connect(device);
-                }
-                break;
-            case REQUEST_ENABLE_BT:
-                // When the request to enable Bluetooth returns
-                if (resultCode == Activity.RESULT_OK) {
-                    // Bluetooth is now enabled, so set up a chat session
-                    setupChat();
-                } else {
-                    // User did not enable Bluetooth or an error occured
-                    Toast.makeText(this, R.string.bt_not_enabled_leaving, Toast.LENGTH_SHORT).show();
-                    finish();
-                }
-        }
-    }
-
-    @Override
-    public void onBackPressed() {
-        Intent i = new Intent(BluetoothTerm.this,HomeActivity.class);
-        i.putExtra("Check","Check");
-        startActivity(i);
-    }
-
-    public void connect(View v) {
-        Intent serverIntent = new Intent(this, DeviceList.class);
-        startActivityForResult(serverIntent, REQUEST_CONNECT_DEVICE);
-    }
-
-    public void discoverable(View v) {
-        ensureDiscoverable();
-    }
-}
