+++ app/src/main/res/layout/activity_dashboard.xml
@@ -37,6 +37,7 @@
                 android:scaleX="0.5"
                 android:background="@drawable/bbluetoth"
                 android:layout_alignParentRight="true"
+                android:onClick="connect"
                 android:layout_centerVertical="true"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content" />
